# Conflicts:
#	src/main/java/austral/ingsis/jjuserservice/controller/UserController.java

+++ src/main/java/austral/ingsis/jjuserservice/service/UserService.java
@@ -1,9 +1,15 @@
 package austral.ingsis.jjuserservice.service;
 
+import austral.ingsis.jjuserservice.dto.ChatUserDto;
+import austral.ingsis.jjuserservice.dto.UpdateUserDto;
 import austral.ingsis.jjuserservice.dto.UserDto;
-import austral.ingsis.jjuserservice.model.User;
+import austral.ingsis.jjuserservice.exception.UserNotFoundException;
+import austral.ingsis.jjuserservice.model.UserDao;
 import austral.ingsis.jjuserservice.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
@@ -11,6 +17,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -18,8 +25,8 @@ public class UserService {
 
     @PersistenceContext
     private EntityManager entityManager; //might not be needed
-    private UserRepository userRepository;
-    private BCryptPasswordEncoder bCryptPasswordEncoder;
+    private final UserRepository userRepository;
+    private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     @Autowired
     public UserService(UserRepository userRepository, BCryptPasswordEncoder encoder) {
@@ -29,21 +36,68 @@ public class UserService {
     }
 
     public List<UserDto> getAllUsers() {
-        List<User> users = this.userRepository.findAll();
-        // communicate with post service to append all posts to object
-        return users.stream().map(User::toUserDto).collect(Collectors.toList());
+        List<UserDao> users = this.userRepository.findAll();
+        //TODO communicate with post service to append all posts to object
+        return users.stream().map(UserDao::toUserDto).collect(Collectors.toList());
     }
 
-    public UserDto saveUser(User user) {
+    public UserDto saveUser(UserDao user) {
+        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
         return this.userRepository.save(user).toUserDto();
     }
 
-    public void deleteUser(String id) {
+    public void deleteUser(Long id) {
+        Optional<UserDao> user = this.userRepository.findById(id);
+        if(user.isEmpty()) {
+            throw new UserNotFoundException("User of id: " + id  + "not found.");
+        }
         this.userRepository.deleteById(id);
     }
 
-    //update
+    public UserDto updateUser(UpdateUserDto dto) {
+        Optional<UserDao> user = this.userRepository.findById(dto.getId());
 
-    //login
+        if(user.isEmpty()) {
+          throw new UserNotFoundException("User of id: " + dto.getId() + "not found.");
+        }
+
+        UserDao toSave = user.get();
+
+        toSave.setEmail(dto.getEmail());
+        toSave.setFirstName(dto.getFirstName());
+        toSave.setLastName(dto.getLastName());
+
+        if(dto.getPassword() != null && !dto.getPassword().isEmpty()) {
+            toSave.setPassword(bCryptPasswordEncoder.encode(dto.getPassword()));
+        }else {
+            toSave.setPassword(user.get().getPassword());
+        }
+
+        return this.userRepository.save(toSave).toUserDto();
+    }
+
+    public UserDto getUserById(Long id) {
+
+        Optional<UserDao> optUser = this.userRepository.findById(id);
+
+       if(optUser.isEmpty()) {
+           throw new UserNotFoundException("User of id: " + id.toString() + "not found.");
+       }
+
+       return optUser.get().toUserDto();
+    }
+
+    public UserDto getUserByUsername(String username) {
+        UserDao user = userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+      return user.toUserDto();
+    }
+
+    public List<UserDto> searchByUsernamePattern(String username, String loggedUser) {
+        return userRepository.findByUsernameStartingWith(username).stream()
+                .filter(found -> !found.getUsername().equals(loggedUser))
+                    .map(UserDao::toUserDto).collect(Collectors.toList());
+//        return null;
+    }
 
 }
